Moved transaction_with_repeatable_read into django_utils
