+++ cvat/apps/dataset_manager/task.py
@@ -28,9 +28,10 @@ from cvat.apps.dataset_manager.util import TmpDirManager, faster_deepcopy
 from cvat.apps.engine import models, serializers
 from cvat.apps.engine.log import DatasetLogManager
 from cvat.apps.engine.plugins import plugin_decorator
-from cvat.apps.engine.utils import av_scan_paths, take_by, transaction_with_repeatable_read
+from cvat.apps.engine.utils import av_scan_paths, take_by
 from cvat.apps.events.handlers import handle_annotations_change
 from cvat.apps.profiler import silk_profile
+from cvat.utils.django_database.contextmanagers import transaction_with_repeatable_read
 from cvat.utils.django_database.utils import add_prefetch_fields, bulk_create, get_cached
 
 dlogger = DatasetLogManager()
