+++ cvat/apps/engine/utils.py
@@ -20,7 +20,7 @@ import traceback
 import urllib.parse
 from collections import defaultdict, namedtuple
 from collections.abc import Callable, Generator, Iterable, Mapping, Sequence
-from contextlib import contextmanager, nullcontext, suppress
+from contextlib import nullcontext, suppress
 from enum import StrEnum, auto
 from itertools import islice
 from multiprocessing import cpu_count
@@ -33,7 +33,6 @@ from av import VideoFrame
 from datumaro.util.os_util import walk
 from django.conf import settings
 from django.core.exceptions import ValidationError
-from django.db import connection, transaction
 from django.db.models import Model
 from django_rq.queues import DjangoRQ
 from django_sendfile import sendfile as _sendfile
@@ -523,15 +522,6 @@ def defaultdict_to_regular(d):
     if isinstance(d, defaultdict):
         d = {k: defaultdict_to_regular(v) for k, v in d.items()}
     return d
-
-
-@contextmanager
-def transaction_with_repeatable_read():
-    with transaction.atomic():
-        if connection.vendor != "sqlite":
-            connection.cursor().execute("SET TRANSACTION ISOLATION LEVEL REPEATABLE READ;")
-            connection.cursor().execute("SET TRANSACTION READ ONLY;")
-        yield
 
 
 def extract_with_patool(archive_path: StrPath, out_dir: StrPath) -> None:
